# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Akasha API",
    "description": "Personal knowledge retrieval.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/files/{id}/enrich": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Ask the language model again for the file's summary and suggested tags\n(e.g. after a failure or a model change). Your own tags are never changed.\nThe result shows up on the file once the background job ran. Needs a\nconfigured language model; limited per user.",
        "operationId": "enrich",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Enrichment queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDetail"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The file is not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No language model is configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionColor": {
        "type": "string",
        "description": "A collection's colour, from the UI's muted palette (never a raw colour).",
        "enum": [
          "sage",
          "sky",
          "ochre",
          "clay",
          "plum",
          "slate"
        ]
      },
...
      "CollectionIcon": {
        "type": "string",
        "description": "A collection's icon.",
        "enum": [
          "folder",
          "book",
          "briefcase",
          "flask",
          "heart",
          "star",
          "archive",
          "receipt",
          "plane",
          "home",
          "graduation",
          "code"
        ]
      },
...
      "CollectionSummary": {
        "type": "object",
        "description": "A collection a file belongs to.",
        "required": [
          "id",
          "name",
          "color",
          "icon"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/CollectionColor"
          },
          "icon": {
            "$ref": "#/components/schemas/CollectionIcon"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "EnrichmentInfo": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "`provider/model` that wrote the summary and tags."
          },
          "status": {
            "$ref": "#/components/schemas/EnrichmentStatus"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "EnrichmentStatus": {
        "type": "string",
        "description": "How the last enrichment (summary and suggested tags) went.",
        "enum": [
          "done",
          "skipped",
          "failed"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "FileDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FileResponse"
          },
          {
            "type": "object",
            "required": [
              "collections"
            ],
            "properties": {
              "collections": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CollectionSummary"
                },
                "description": "The collections the file is in, by name."
              },
              "processing": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ProcessingJob",
                    "description": "Latest processing job (extraction or embedding), if one was ever queued."
                  }
                ]
              }
            }
          }
        ],
        "description": "One file plus the state of its latest processing job."
      },
...
      "FileResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "mime_type",
          "size_bytes",
          "content_hash",
          "status",
          "is_pinned",
          "tags",
          "auto_tags",
          "created_at",
          "updated_at",
          "open_count"
        ],
        "properties": {
          "auto_tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags suggested by the language model, minus any already in `tags`.\nFiltering and searching by tag match both lists."
          },
          "content_hash": {
            "type": "string",
            "description": "SHA-256 of the contents, lowercase hex."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enrichment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EnrichmentInfo",
                "description": "State of the model-written summary and tags; `null`: not enriched (yet)."
              }
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why processing failed, when `status` is `failed`."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_pinned": {
            "type": "boolean"
          },
          "last_opened_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When you last opened the file's page (`POST /files/{id}/open`)."
          },
          "mime_type": {
            "type": "string",
            "description": "Detected from the contents, not taken from the client."
          },
          "name": {
            "type": "string",
            "description": "Sanitised original filename."
          },
          "open_count": {
            "type": "integer",
            "format": "int32"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "$ref": "#/components/schemas/FileStatus"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ],
            "description": "A short model-written description, once the file was enriched."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The user's own tags (never changed by the model)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FileStatus": {
        "type": "string",
        "description": "Processing state of a file.",
        "enum": [
          "pending",
          "processing",
          "ready",
          "failed"
        ]
      },
...
      "JobState": {
        "type": "string",
        "description": "State of a background job.",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "dead"
        ]
      },
...
      "ProcessingJob": {
        "type": "object",
        "required": [
          "stage",
          "state",
          "attempts",
          "max_attempts",
          "updated_at"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Attempts started so far."
          },
          "max_attempts": {
            "type": "integer",
            "format": "int32"
          },
          "next_attempt_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When it will next be tried (`queued` or `failed` only)."
          },
          "stage": {
            "$ref": "#/components/schemas/ProcessingStage"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProcessingStage": {
        "type": "string",
        "description": "Which processing step a job performs.",
        "enum": [
          "extract",
          "embed"
        ]
      },
...
    },
    "securitySchemes": {
      "api_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal API token (akasha_pat_…)"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "akasha_session"
      }
    }
  }
}